+++ frontend/src/metabase/dashboard/components/DashCard/DashCardCardParameterMapper/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DashCardCardParameterMapper";
