+++ frontend/src/metabase/dashboard/components/DashCard/DashCardParameterMapper.jsx
@@ -3,7 +3,7 @@ import React from "react";
 import { t } from "ttag";
 
 import { color } from "metabase/lib/colors";
-import DashCardCardParameterMapper from "../components/DashCardCardParameterMapper";
+import DashCardCardParameterMapper from "./DashCardCardParameterMapper";
 
 const DashCardParameterMapper = ({ dashcard, isMobile }) => (
   <div className="relative flex-full flex flex-column layout-centered">
